+++ tests/test_wl_keyboard.py
@@ -3,7 +3,7 @@ import time
 
 import wayland as wl
 
-wayland = wl.initialise()
+wayland = wl.client.initialise()
 
 
 def test_keyboard():
@@ -34,8 +34,7 @@ def test_keyboard():
 
     start = time.time()
     while not keymap and time.time() < start + 3:
-        time.sleep(0.1)
-        wayland.process_messages()
+        wayland.wl_display.dispatch_timeout(0.2)
 
     # If we don't have a keyboard, don't test it
     if not have_keyboard: