+++ tests/test_wl_registry.py
@@ -1,7 +1,6 @@
 import wayland as wl
-from tests.utils import process_messages
 
-wayland = wl.initialise()
+wayland = wl.client.initialise()
 
 
 def test_get_registry():
@@ -16,7 +15,7 @@ def test_get_registry():
     # Hook the event to get the registry results
     wayland.wl_registry.events.global_ += on_wl_registry_global
     wayland.wl_display.get_registry()
-    process_messages(wayland)
+    wayland.wl_display.dispatch_timeout(2.0)
 
     # Check we got some interfaces we should have
     for proto in protocols: