+++ CHANGELOG.md
@@ -1,5 +1,21 @@
 ### Changelog
 
+### v1.0.0 (xxth xxx 2025)
+- Event dispatching changed to more closely align with the pattern in `libwayland-client`, using dispatch and dispatch_pending for blocking and non-blocking event dispatching. (see the below breaking API changes)
+- Event handling changed to correctly support asynchronous event processing across multiple threads, rather than the previous synchronous, single threaded event queue.
+- All non-wayland protocol functionality introduced by `python-wayland` moved to the namespace `wayland.client`
+- The `wayland` package namespace cleaned up so it only contains wayland interfaces and the `client` module.
+- Library API documentation added.
+- The following critical breaking API change were made:
+- `wayland.process_messages()` removed.
+- `wayland.wl_display.dispatch()` added.
+- `wayland.wl_display.dispatch_timeout()` added.
+- `wayland.wl_display.dispatch_pending()` added.
+- The following breaking API changes were made, although these methods not required for the normal use of this library:
+- `wayland.initialise()` changed to `wayland.client.initialise()`
+- `wayland.get_package_root()` changed to `wayland.client.get_package_root()`
+- `wayland.is_wayland` changed to `wayland.client.is_wayland()` (note it became a function too)
+
 ### v0.7.1 (28th May 2025)
 - Remove dependency on requests library.
 - Lint fixes for unit tests.